requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-common-helpers==0.9.2
django-filter==2.1.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
django-crispy-forms==1.7.2
future==0.17.1
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
kombu==4.5.0
MarkupSafe==1.1.1

openapi-codec==1.3.2
pi==0.1.2
Pillow==6.0.0
PyJWT==1.7.1
pytz==2019.1

requests==2.21.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.3
vine==1.3.0
